COMP: RunLength signed / unsigned comparison
To address:

 c:\jenkins\workspace\itkgerritwindows\itk-src\modules\remote\texturefeatures\include\itkRunLengthTextureFeaturesImageFilter.hxx(404):
 warning C4018: '<': signed/unsigned mismatch

+++ include/itkRunLengthTextureFeaturesImageFilter.hxx
@@ -398,10 +398,10 @@ RunLengthTextureFeaturesImageFilter<TInputImage, TOutputImage>
     }
   offsetDistance = std::sqrt(offsetDistance);
 
-  int offsetDistanceBin = (int)(( offsetDistance*pixelDistance - m_MinDistance)/
+  int offsetDistanceBin = static_cast< int>(( offsetDistance*pixelDistance - m_MinDistance)/
           ( (m_MaxDistance - m_MinDistance) / (float)m_NumberOfBinsPerAxis ));
 
-  if (offsetDistanceBin < m_NumberOfBinsPerAxis)
+  if (offsetDistanceBin < static_cast< int >( m_NumberOfBinsPerAxis ))
     {
     totalNumberOfRuns++;
     hist[curentInNeighborhoodPixelIntensity][offsetDistanceBin]++;
